minimum order update

+++ accounts/models.py
@@ -378,7 +378,7 @@ class Restaurant(User):
     restaurantLogo=models.ImageField(upload_to='proven/')
     commercialRecordImage=models.ImageField(upload_to='proven/')
     restaurantStatus=models.BooleanField(default=True)
-    minimumOrder=models.PositiveSmallIntegerField(default=1)
+    minimumOrder=models.DecimalField(max_digits=8, decimal_places=2,default=1.0)
     
     
     def save(self, *args, **kwargs):
